Add SRv6 5 node docker demo
This demo is intended to test SRv6 implementation. There are
two options how to run the demo. First one is to use VPP only
configuration and the second one uses honeycomb to configure
SRv6 on nodes.

For detailed information about the demo read documentation in
./asciidoc/Readme.adoc file. Follow the instructions to run
the demo and use provided commands to verify the operation
of VPP nodes.

Change-Id: I7d17db0096cc48b06e9148e7c78faaf9d7b8a97d
Signed-off-by: Michal Cmarada <michal.cmarada@pantheon.tech>