Add configuration for mtu in docker examples
VPP is using mtu=9000 for interfaces that it creates.

This fixes the mtu value for veth interfaces on linux side
so they match the value configured in VPP for all veth interfaces
in docker examples. Originally veth interfaces use smaller value
since Linux automatically sets their MTU to be the smallest MTU
of their slave ports. For VPP this means if a bigger frame is sent
outside to the veth interface it will be dropped.

Change-Id: I41117490ec80a0b2d2d2897493c2690f92766734
Signed-off-by: Michal Cmarada <michal.cmarada@pantheon.tech>